Splits it up
