+++ z_order.rb
@@ -1,0 +1,3 @@
+module ZOrder
+  Background, Enemys, Player, Projectile, UI = *0..4
+end
